+++ frontend/src/routes/+page.svelte
@@ -1,64 +1,4 @@
 <script lang="ts">
-  // import axios from 'axios';
-  import { onMount, createEventDispatcher } from 'svelte';
-
-  let daysOfWeek = [
-    { name: 'Monday', selected: false },
-    { name: 'Tuesday', selected: false },
-    { name: 'Wednesday', selected: false },
-    { name: 'Thursday', selected: false },
-    { name: 'Friday', selected: false },
-    { name: 'Saturday', selected: false },
-    { name: 'Sunday', selected: false }
-  ];
-
-  let selectAll = false;
-
-  const toggleAll = () => {
-    selectAll = !selectAll;
-    daysOfWeek = daysOfWeek.map((day) => ({ ...day, selected: selectAll }));
-  };
-
-  let goalDate: string;
-  let goalTime: string;
-
-  const dispatch = createEventDispatcher();
-
-  function handleClick() {
-    dispatch('save');
-    // add your saving logic here
-  }
-
-  let loadingGenerate = false;
-  let loadingRegenerate = false;
-
-  const handleClickGenerate = async () => {
-    loadingGenerate = true;
-    // Simulate an API call delay
-    await new Promise((resolve) => setTimeout(resolve, 2000));
-    loadingGenerate = false;
-  };
-
-  const handleClickRegenerate = async () => {
-    loadingRegenerate = true;
-    // Simulate an API call delay
-    await new Promise((resolve) => setTimeout(resolve, 2000));
-    loadingRegenerate = false;
-  };
-
-  onMount(async () => {
-    // Fetch data from API
-    // const response = await axios.get('your-api-endpoint');
-    // const data = response.data;
-    // Assign the fetched data to variables or update the necessary parts of the component
-    // goalDate = data.goalDate;
-    // goalTime = data.goalTime;
-  });
-
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  function regenerate(event: MouseEvent & { currentTarget: EventTarget & HTMLButtonElement }) {
-    throw new Error('Function not implemented.');
-  }
 </script>
 
 <div class="hero min-h-screen bg-base-200 mb-10">
